requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.10
fastapi-cli==0.0.7
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.5
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typer==0.15.2
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
